requirements.txt
# Core AI
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0

# Atlassian
atlassian-python-api>=3.41.0

# Snowflake
snowflake-connector-python>=3.10.0

# Data validation
pydantic>=2.7.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Scheduling & watching
apscheduler>=3.10.0
watchdog>=4.0.0

# ER Diagrams
mermaid-py>=0.5.0

# CLI & output
rich>=13.7.0
typer>=0.12.0

# Web dashboard
fastapi>=0.111.0
uvicorn>=0.29.0

# Utilities
httpx>=0.27.0
tenacity>=8.3.0

# Vector store for semantic search
chromadb>=0.5.0

# Multi-LLM support
anthropic>=0.25.0
google-generativeai>=0.5.0

# Async support
aiofiles>=23.2.0
asyncio>=3.4.3

# GraphQL API
strawberry-graphql>=0.227.0

# WebSocket support
websockets>=12.0

# Caching
cachetools>=5.3.0

# Background tasks
celery>=5.4.0
redis>=5.0.0

# Security
python-jose>=3.3.0
passlib>=1.7.0
bcrypt>=4.1.0

# Metrics & monitoring
prometheus-client>=0.20.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# Documentation
mkdocs>=1.6.0
mkdocs-material>=9.5.0
